requirements.txt

torch==2.4.1+cu121
torchvision==0.19.1+cu121
torchaudio==2.4.1+cu121
torchmetrics==1.4.1
pytorch-lightning==2.2.4
scikit-learn==1.3.2
xgboost==2.1.1
lightning-utilities==0.11.7
shap==0.44.1
transformers==4.44.2
statsforecast==1.7.6
statsmodels==0.14.1
pandas==1.5.3
numpy==1.24.4
scipy==1.10.1
matplotlib==3.7.5
seaborn==0.13.2
joblib==1.4.2
numba==0.58.1
tensorboard==2.14.0
tensorboardx==2.6.2.2
huggingface-hub==0.24.6
aiohttp==3.10.5
requests==2.32.3
fsspec==2024.9.0
cloudpickle==3.0.0
jsonschema==4.19.2
click==8.1.7
u8darts==0.29.0
